refactor(core): move the geo-metadata layer out of core/common.py
The Arrow-table geo block and the written-file geo block become two modules
under the write path. Pure relocation; every name is still importable from
core.common.
