+++ pyproject.toml
@@ -214,9 +214,19 @@ ignore_missing_imports = true
 # Per-module overrides for files with known type issues.
 # These suppress errors in complex modules until they are incrementally fixed.
 # TODO: Remove overrides as type annotations are added to each module.
+#
+# `arrow_geo_metadata` and `geo_metadata_repair` were carved out of
+# `geoparquet_io.core.common`, which is on this list: their seven errors were
+# always suppressed, by their old address. Listing them relocates that
+# suppression rather than granting a new one, so the split stays a pure move.
+# Typing them out is tracked separately -- the errors are real (a
+# `compression_level: int | None` reaching a field typed `int`, an
+# `effective_version: str | None` used as a dict key), not annotation noise.
 [[tool.mypy.overrides]]
 module = [
     "geoparquet_io.api.check",
+    "geoparquet_io.core.arrow_geo_metadata",
+    "geoparquet_io.core.geo_metadata_repair",
     "geoparquet_io.api.ops",
     "geoparquet_io.api.table",
     "geoparquet_io.benchmarks.profile_report",
